+++ src/main/java/iris/command/ExitCommand.java
@@ -1,4 +1,8 @@
-package src.main.java.command;
+package iris.command;
+
+import iris.TaskList;
+import iris.Ui;
+import iris.TaskStore;
 
 public class ExitCommand extends Command{
     @Override
