+++ src/main/java/iris/command/FilterCommand.java
@@ -1,4 +1,10 @@
-package src.main.java.command;
+package iris.command;
+
+import iris.TaskList;
+import iris.Ui;
+import iris.TaskStore;
+import iris.exception.DateTimeException;
+import iris.exception.IrisException;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
