+++ src/main/java/iris/command/ResetCommand.java
@@ -1,4 +1,9 @@
-package src.main.java.command;
+package iris.command;
+
+import iris.TaskList;
+import iris.Ui;
+import iris.exception.IrisException;
+import iris.TaskStore;
 
 public class ResetCommand extends Command {
     @Override
