+++ src/main/java/iris/exception/MissingFieldException.java
@@ -1,4 +1,4 @@
-package src.main.java.exception;
+package iris.exception;
 
 public class MissingFieldException extends IrisException {
     public MissingFieldException(String field) {
