+++ src/main/java/iris/exception/NoTaskException.java
@@ -1,4 +1,4 @@
-package src.main.java.exception;
+package iris.exception;
 
 public class NoTaskException extends IrisException {
     public NoTaskException() {
