+++ src/main/java/iris/task/Todo.java
@@ -1,7 +1,7 @@
-package src.main.java.task;
+package iris.task;
 
 public class Todo extends Task{
-    Todo(String name) {
+    public Todo(String name) {
         super(name);
     }
